restructuring routes to be a bit more logical and not use conditional rendering in root.tsx
